requirements.txt
asgiref==3.8.1
dj-database-url==2.3.0
Django==5.1.6
django-cors-headers==4.7.0
djangorestframework==3.15.2
git-filter-repo==2.47.0
packaging==24.2
psycopg2==2.9.10
psycopg2-binary==2.9.10
python-dotenv==1.0.1
setuptools==75.8.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
waitress==3.0.2
wheel==0.45.1
whitenoise==6.9.0
pytesseract==0.3.13
nltk==3.9.1
ollama==0.1.14
pdfplumber==0.10.3